+++ src/detection/terminalshell/terminalshell.c
@@ -155,3 +155,41 @@ bool fftsGetShellVersion(FFstrbuf* exe, const char* exeName, FFstrbuf* version)
         ok = false;
     return ok;
 }
+
+bool fftsGetTerminalVersion(FFstrbuf* processName, FF_UNUSED_PARAM FFstrbuf* exe, FFstrbuf* version)
+{
+    #ifdef __ANDROID__
+
+    if(ffStrbufEqualS(processName, "Termux"))
+    {
+        ffStrbufSetS(version, getenv("TERMUX_VERSION"));
+        return true;
+    }
+
+    #endif
+
+    const char* termProgramVersion = getenv("TERM_PROGRAM_VERSION");
+    if(termProgramVersion)
+    {
+        const char* termProgram = getenv("TERM_PROGRAM");
+        if(termProgram)
+        {
+            if(ffStrbufStartsWithIgnCaseS(processName, termProgram) || // processName ends with `.exe` on Windows
+                (strcmp(termProgram, "vscode") == 0 && ffStrbufStartsWithIgnCaseS(processName, "code"))
+            ) {
+                ffStrbufSetS(version, termProgramVersion);
+                return true;
+            }
+        }
+    }
+
+    #ifdef _WIN32
+
+    return getFileVersion(exe->chars, version);
+
+    #else
+
+    return false;
+
+    #endif
+}